+++ integrate/integrate.h
@@ -56,22 +56,22 @@ __float128 Integrate_Simpson_f128(__float128 func(__float128,void*),
 Gauss-Legendre integrate
 ===============================================================================
 */
-bool Integrate_Gauss_Legendre_x(uint64_t n_int_, double *x_int);
-bool Integrate_Gauss_Legendre_w(uint64_t n_int, double *x_int, double *w_int);
+bool Integrate_Gauss_Legendre_x(int32_t n_int_, double *x_int);
+bool Integrate_Gauss_Legendre_w(int32_t n_int, double *x_int, double *w_int);
 double Integrate_Gauss_Legendre_via_xw(double func(double, void *), double xmin, double xmax,
-                                       uint64_t n_int, void *other_param,
+                                       int32_t n_int, void *other_param,
                                        double *x_int, double *w_int);
 double Integrate_Gauss_Legendre(double func(double, void *), double xmin, double xmax,
-                                uint64_t n_int, void *other_param);
+                                int32_t n_int, void *other_param);
 
-/* uint32_t & float */
-bool Integrate_Gauss_Legendre_x_f32(uint32_t n_int_, float *x_int);
-bool Integrate_Gauss_Legendre_w_f32(uint32_t n_int, float *x_int, float *w_int);
+/* float */
+bool Integrate_Gauss_Legendre_x_f32(int32_t n_int_, float *x_int);
+bool Integrate_Gauss_Legendre_w_f32(int32_t n_int, float *x_int, float *w_int);
 float Integrate_Gauss_Legendre_via_xw_f32(float func(float, void *), float x1, float x2,
-                                          uint32_t n_int, void *other_param,
+                                          int32_t n_int, void *other_param,
                                           float *x_int, float *w_int);
 float Integrate_Gauss_Legendre_f32(float func(float, void *), float x1, float x2,
-                                   uint32_t n_int, void *other_param);
+                                   int32_t n_int, void *other_param);
 
 /*
 ===============================================================================
